pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-dipg-safety"
version = "0.1.0"
description = "DIPG Safety Environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.1.0",  # The core dependency
    "fastapi>=0.109.1",
    "uvicorn[standard]>=0.24.0",  # Changed from == to >= to allow mcp to use newer version
    "requests>=2.31.0",
    "wsproto==1.0.0",
    "gunicorn==22.0.0",
    "datasets",
    "starlette>=0.36.3",
    "idna>=3.7",
    "mcp>=1.0.0",  # For MCP server functionality
    "fastmcp>=0.5.0",  # For FastMCP-based MCP server
    "google-adk>=1.0.0",  # For ADK agent
    "a2a-sdk>=0.1.0",  # For A2A protocol
    "litellm>=1.0.0",  # For LiteLLM model support
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ipykernel>=6.29.5",
]
visualization = [
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scipy>=1.11.0",
    "pandas>=2.1.0",
    "reportlab>=4.0.0",  # For PDF generation
]

# This creates a command 'dipg-server' you can run from the terminal
[project.scripts]
dipg-server = "med_safety_gym.app:main"

[tool.setuptools]
packages = ["med_safety_gym", "tests"]  # Includes your gym and test modules
